The Chuckle Factory - V1.0.1
Minor Update:
Added P tags and used CSS to align it into Center.

There could be Minor Updates here and there for fix bugs in the Webpage.
If any further error found they will be Debugged and mentioned in the Source Control and they will fixed and mentioned.
Also, how I come up with a solution and what I did to the code in able to fix that error.
There may be few updates here and there if any further errors located.

**Additional Quality of LIfe Update**